check if your card has been leaked
This project was developed as part of an academic activity at college, with the aim of demonstrating how people can fall victim to social engineering scams. The code was created for educational purposes, raising awareness about the risks associated with fraudulent practices on the internet.

Attention: This project is not intended to cause harm or deceive anyone, and is merely a demonstration of the type of technique used in scams.